Module brontes_database_libmdbx::tables

Structs

AddressToFactory

address -> factory Takes Address as a key and returns StaticBindingsDb.

AddressToProtocol

Address -> Static protocol enum Takes Address as a key and returns StaticBindingsDb.

AddressToTokens

Address -> tokens in pool Takes Address as a key and returns PoolTokens.

CexPrice

block num -> cex prices Takes u64 as a key and returns CexPriceMap.

DexPrice

block number concat tx idx -> cex quotes Takes TxHash as a key and returns DexQuoteWithIndex.

Metadata

block num -> metadata Takes u64 as a key and returns MetadataInner.

MevBlocks

block number -> mev block with classified mev Takes u64 as a key and returns MevBlockWithClassified.

PoolCreationBlocks

block number -> pools created in block Takes u64 as a key and returns PoolsLibmdbx.

SubGraphs

pair -> Vec<(block_number, entry)> Takes Pair as a key and returns SubGraphsEntry.

TokenDecimals

token address -> decimals Takes Address as a key and returns u8.

Enums

Tables

Constants

NUM_TABLES

Traits

IntoTableKey